e09

+++ e09/E09.java
@@ -4,7 +4,7 @@ class E09 {
     public static void main(String[] args) {
         Scanner scanner = new Scanner(System.in);
 
-        int total = 0;
+        double total = 0;
         int count = 0;
 
         while (true) {
